Renamed Stats in estimator
