+++ mlcolvar/core/__init__.py
@@ -1,7 +1,6 @@
-__all__ = ["loss", "nn", "transform", "stats"]
+__all__ = ["loss", "nn", "transform", "estimators"]
 
 from .loss import *
 from .nn import *
 from .transform import *
-from .stats import *
-from .operator_learning import *
+from .estimators import *
