+++ mlcolvar/core/estimators/__init__.py
@@ -1,0 +1,7 @@
+__all__ = ["Estimator", "PCA", "LDA", "TICA", "Generator"]
+
+from .estimator import *
+from .pca import *
+from .lda import *
+from .tica import *
+from .generator import *
