+++ mlcolvar/core/estimators/utils.py
