--- mlcolvar/core/operator_learning/__init__.py
@@ -1,3 +1,0 @@
-__all__ = ["Generator"]
-
-from .generator import *
