--- mlcolvar/core/stats/__init__.py
@@ -1,6 +1,0 @@
-__all__ = ["Stats", "PCA", "LDA", "TICA"]
-
-from .stats import *
-from .pca import *
-from .lda import *
-from .tica import *
