+++ mlcolvar/cvs/generator/__init__.py
@@ -1,4 +1,4 @@
-__all__ = ["DeepGenerator", "compute_eigenfunctions", "forecast_state_occupation"]
+__all__ = ["DeepGenerator", "forecast_state_occupation"]
 
 from .generator import *
 from .utils import *
