+++ mlcolvar/tests/test_core_stats_lda.py
@@ -1,6 +1,6 @@
 import pytest
 
-from mlcolvar.core.stats.lda import test_lda
+from mlcolvar.core.estimators.lda import test_lda
 
 if __name__ == "__main__":
     test_lda()
