+++ mlcolvar/tests/test_core_stats_pca.py
@@ -1,6 +1,6 @@
 import pytest
 
-from mlcolvar.core.stats.pca import test_pca
+from mlcolvar.core.estimators.pca import test_pca
 
 if __name__ == "__main__":
     test_pca()
