+++ mlcolvar/tests/test_core_stats_tica.py
@@ -1,6 +1,6 @@
 import pytest
 
-from mlcolvar.core.stats.tica import test_tica  # ,test_reduced_rank_tica
+from mlcolvar.core.estimators.tica import test_tica  # ,test_reduced_rank_tica
 
 if __name__ == "__main__":
     test_tica()
