+++ mlcolvar/tests/test_cvs_generator.py
@@ -1,7 +1,5 @@
 from mlcolvar.cvs.generator.generator import test_generator, test_generator_runtime_cell_training
-from mlcolvar.cvs.generator.utils import test_forecast_state_occupation
 
 if __name__ == "__main__":
     test_generator()
     test_generator_runtime_cell_training()
-    test_forecast_state_occupation()
